requirements.txt
streamlit>=1.37.0
qrcode[pil]>=7.4.2
Pillow>=10.3.0
python-dotenv>=1.0.1
reportlab>=4.0.0
